+++ warp/tests/unittest_suites.py
@@ -132,6 +132,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
     from warp.tests.test_constant_precision import TestConstantPrecision
     from warp.tests.test_context import TestContext
     from warp.tests.test_copy import TestCopy
+    from warp.tests.test_cpu_precompiled_headers import TestCpuPrecompiledHeaders
     from warp.tests.test_ctypes import TestCTypes
     from warp.tests.test_dense import TestDense
     from warp.tests.test_devices import TestDevices
@@ -233,6 +234,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
         TestConstantPrecision,
         TestContext,
         TestCopy,
+        TestCpuPrecompiledHeaders,
         TestCTypes,
         TestCudaArchSuffix,
         TestDense,